requirements.txt
Django==4.1
django-admin-sortable2==2.1.2
environs==9.5.0
Pillow==9.2.0